--- a/benchmark/benchmarks.jl
+++ b/benchmark/benchmarks.jl
@@ -1,8 +1,5 @@
 using BenchmarkTools
 
-# Default benchmark parameters
-BenchmarkTools.DEFAULT_PARAMETERS.seconds = 2.5
-
 # Benchmark suite modules
 const SUITE_MODULES = Dict("gcp" => :BenchmarkGCP, "mttkrp" => :BenchmarkMTTKRP)