--- a/benchmark/run.jl
+++ b/benchmark/run.jl
@@ -72,7 +72,7 @@ end
 ### Build report
 report = sprint(export_markdown, results)
 report *= "\n\n" * GCPBenchmarkUtils.export_mttkrp_sweep(results)
-report = collapsible_details(report)
+report = GCPBenchmarkUtils.collapsible_details(report)
 
 ### Save report
 write(joinpath(@__DIR__, "report.md"), report)