--- a/benchmark/suites/mttkrp.jl
+++ b/benchmark/suites/mttkrp.jl
@@ -40,8 +40,11 @@ for SETUP in SETUPS
     Random.seed!(0)
     X = randn(SETUP.size)
     U = [randn(In, SETUP.rank) for In in SETUP.size]
-    SUITE["size=$(SETUP.size), rank=$(SETUP.rank), mode=$(SETUP.mode)"] =
-        @benchmarkable GCPDecompositions.mttkrp($X, $U, $(SETUP.mode))
+    SUITE["size=$(SETUP.size), rank=$(SETUP.rank), mode=$(SETUP.mode)"] = @benchmarkable(
+        GCPDecompositions.mttkrp($X, $U, $(SETUP.mode)),
+        seconds = 2,
+        samples = 5,
+    )
 end
 
 end