Make some additional enhancements:

+ Add another imbalanced tensor (with dimensions reversed)
+ Set benchmark parameters on case-by-case basis
  + e.g., `mttkrp` runs should be faster than full `gcp` runs so should get less time
  + also set the number of samples for `mttkrp` (that way "faster methods = faster benchmarking")
+ Make details collapsible
  + addresses item in https://github.com/dahong67/GCPDecompositions.jl/pull/30#issuecomment-1811370872
+ Replace path to package with package name
  + Don't want to leak that info when we post on PRs/etc.