@dahong67 2023-11-15
With this PR, running
```
julia benchmark/run.jl --suite mttkrp --compare 11d1d29
```
produces the following report.

# Benchmark Report for `GCPDecompositions`

## Job Properties

<details>

* Time of benchmarks:
    - Target: 14 Nov 2023 - 19:36
    - Baseline: 14 Nov 2023 - 19:39
* Package commits:
    - Target: 11d1d2
    - Baseline: 11d1d2
* Julia commits:
    - Target: bed2cd
    - Baseline: bed2cd
* Julia command flags:
    - Target: None
    - Baseline: None
* Environment variables:
    - Target: `GCP_BENCHMARK_SUITES => mttkrp`
    - Baseline: `GCP_BENCHMARK_SUITES => mttkrp`

</details>


## Results

<details>

A ratio greater than `1.0` denotes a possible regression (marked with :x:), while a ratio less
than `1.0` denotes a possible improvement (marked with :white_check_mark:). Only significant results - results
that indicate possible regressions or improvements - are shown below (thus, an empty table means that all
benchmark results remained invariant between builds).

| ID                                                     | time ratio                   | memory ratio |
|--------------------------------------------------------|------------------------------|--------------|
| `["mttkrp", "size=(150, 150, 150), rank=150, mode=2"]` | 0.94 (5%) :white_check_mark: |   1.00 (1%)  |

</details>


## Benchmark Group List

<details>

Here's a list of all the benchmark groups executed by this job:

- `["mttkrp"]`

</details>


## Julia versioninfo


<details>

### Target
```
Julia Version 1.9.3
Commit bed2cd540a1 (2023-08-24 14:43 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: macOS (arm64-apple-darwin22.4.0)
  uname: Darwin 23.0.0 Darwin Kernel Version 23.0.0: Fri Sep 15 14:41:43 PDT 2023; root:xnu-10002.1.13~1/RELEASE_ARM64_T6000 arm64 arm
  CPU: Apple M1 Max: 
                 speed         user         nice          sys         idle          irq
       #1-10  2400 MHz    2026224 s          0 s     751946 s   30863090 s          0 s
  Memory: 64.0 GB (22852.984375 MB free)
  Uptime: 347696.0 sec
  Load Avg:  2.58544921875  2.9931640625  2.6298828125
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, apple-m1)
  Threads: 1 on 8 virtual cores
```

### Baseline
```
Julia Version 1.9.3
Commit bed2cd540a1 (2023-08-24 14:43 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: macOS (arm64-apple-darwin22.4.0)
  uname: Darwin 23.0.0 Darwin Kernel Version 23.0.0: Fri Sep 15 14:41:43 PDT 2023; root:xnu-10002.1.13~1/RELEASE_ARM64_T6000 arm64 arm
  CPU: Apple M1 Max: 
                 speed         user         nice          sys         idle          irq
       #1-10  2400 MHz    2029223 s          0 s     752623 s   30879892 s          0 s
  Memory: 64.0 GB (23463.0 MB free)
  Uptime: 347901.0 sec
  Load Avg:  2.744140625  2.91259765625  2.673828125
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, apple-m1)
  Threads: 1 on 8 virtual cores
```

</details>


# MTTKRP benchmark plots

## Runtime vs. size (for square tensors)

<details>

Below are plots showing the runtime in miliseconds of MTTKRP as a function of the size of the square tensor, for varying ranks and modes:
<table>
<tr>
<th></th>
<th>ndims = 3, rank = 50, mode = 1</th>
<th>ndims = 3, rank = 50, mode = 2</th>
<th>ndims = 3, rank = 50, mode = 3</th>
<th>ndims = 3, rank = 100, mode = 1</th>
<th>ndims = 3, rank = 100, mode = 2</th>
<th>ndims = 3, rank = 100, mode = 3</th>
<th>ndims = 3, rank = 150, mode = 1</th>
<th>ndims = 3, rank = 150, mode = 2</th>
<th>ndims = 3, rank = 150, mode = 3</th>
<th>ndims = 3, rank = 200, mode = 1</th>
<th>ndims = 3, rank = 200, mode = 2</th>
<th>ndims = 3, rank = 200, mode = 3</th>
<th>ndims = 3, rank = 250, mode = 1</th>
<th>ndims = 3, rank = 250, mode = 2</th>
<th>ndims = 3, rank = 250, mode = 3</th>
<th>ndims = 3, rank = 300, mode = 1</th>
<th>ndims = 3, rank = 300, mode = 2</th>
<th>ndims = 3, rank = 300, mode = 3</th>
</tr>
<tr>
<th>Target</th>
<td>

```
               ndims = 3, rank = 50, mode = 1  
              ┌──────────────────────────────┐ 
          200 │                             .│ 
              │                            .'│ 
              │                           :  │ 
              │                          :   │ 
Time (ms)     │                        .'    │ 
              │                       .'     │ 
              │                      .'      │ 
              │                   .''        │ 
              │                .''           │ 
            0 │       ....'''''              │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 50, mode = 2  
              ┌──────────────────────────────┐ 
          200 │                              │ 
              │                              │ 
              │                             :│ 
              │                           .' │ 
Time (ms)     │                         .'   │ 
              │                        :     │ 
              │                      .'      │ 
              │                   .''        │ 
              │                .''           │ 
            0 │       ....'''''              │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 50, mode = 3  
              ┌──────────────────────────────┐ 
          200 │                              │ 
              │                              │ 
              │                             .│ 
              │                           .' │ 
Time (ms)     │                         .'   │ 
              │                        .'    │ 
              │                      .'      │ 
              │                   ..'        │ 
              │                .''           │ 
            0 │       ....'''''              │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 100, mode = 1 
              ┌──────────────────────────────┐ 
          400 │                              │ 
              │                             :│ 
              │                           .' │ 
              │                          .'  │ 
Time (ms)     │                         :    │ 
              │                       .'     │ 
              │                      .'      │ 
              │                   ..'        │ 
              │                .''           │ 
            0 │       ....'''''              │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 100, mode = 2 
              ┌──────────────────────────────┐ 
          300 │                             :│ 
              │                           .' │ 
              │                          .'  │ 
              │                         :    │ 
Time (ms)     │                       .'     │ 
              │                      .'      │ 
              │                    .'        │ 
              │                 ..'          │ 
              │             ...'             │ 
            0 │       ...'''                 │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 100, mode = 3 
              ┌──────────────────────────────┐ 
          300 │                             :│ 
              │                           .' │ 
              │                          .'  │ 
              │                         :    │ 
Time (ms)     │                       .'     │ 
              │                      .'      │ 
              │                    ..'       │ 
              │                  .'          │ 
              │             ...''            │ 
            0 │       ...'''                 │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 150, mode = 1 
              ┌──────────────────────────────┐ 
          600 │                              │ 
              │                            .'│ 
              │                           .' │ 
              │                          :   │ 
Time (ms)     │                         :    │ 
              │                       .'     │ 
              │                      .'      │ 
              │                   ..'        │ 
              │                .''           │ 
            0 │       ....'''''              │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 150, mode = 2 
              ┌──────────────────────────────┐ 
          500 │                              │ 
              │                            .'│ 
              │                           .' │ 
              │                          :   │ 
Time (ms)     │                        .'    │ 
              │                       :      │ 
              │                     .'       │ 
              │                  .''         │ 
              │              ...'            │ 
            0 │       ...''''                │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 150, mode = 3 
              ┌──────────────────────────────┐ 
          500 │                              │ 
              │                             :│ 
              │                           .' │ 
              │                          :   │ 
Time (ms)     │                        .'    │ 
              │                       :      │ 
              │                     .'       │ 
              │                  .''         │ 
              │              ...'            │ 
            0 │       ...''''                │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 200, mode = 1 
              ┌──────────────────────────────┐ 
          800 │                              │ 
              │                             :│ 
              │                           .' │ 
              │                          .'  │ 
Time (ms)     │                         :    │ 
              │                       .'     │ 
              │                      .'      │ 
              │                   ..'        │ 
              │                .''           │ 
            0 │       ....'''''              │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 200, mode = 2 
              ┌──────────────────────────────┐ 
          600 │                             :│ 
              │                           .' │ 
              │                          .'  │ 
              │                         :    │ 
Time (ms)     │                       .'     │ 
              │                      .'      │ 
              │                    .'        │ 
              │                 ..'          │ 
              │             ...'             │ 
            0 │       ...'''                 │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 200, mode = 3 
              ┌──────────────────────────────┐ 
          600 │                             :│ 
              │                           .' │ 
              │                          .'  │ 
              │                         :    │ 
Time (ms)     │                       .'     │ 
              │                      .'      │ 
              │                    .''       │ 
              │                  .'          │ 
              │             ...''            │ 
            0 │       ...'''                 │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 250, mode = 1 
              ┌──────────────────────────────┐ 
          900 │                             :│ 
              │                           .' │ 
              │                          .'  │ 
              │                         .'   │ 
Time (ms)     │                        :     │ 
              │                       :      │ 
              │                     .:       │ 
              │                  ..'         │ 
              │               ..'            │ 
            0 │       ....''''               │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 250, mode = 2 
              ┌──────────────────────────────┐ 
          800 │                              │ 
              │                             :│ 
              │                           .' │ 
              │                          :   │ 
Time (ms)     │                        .'    │ 
              │                       :      │ 
              │                     .'       │ 
              │                  .''         │ 
              │              ..''            │ 
            0 │       ...''''                │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 250, mode = 3 
              ┌──────────────────────────────┐ 
          800 │                             .│ 
              │                            .'│ 
              │                           :  │ 
              │                         .'   │ 
Time (ms)     │                        .'    │ 
              │                       :      │ 
              │                     .'       │ 
              │                  .''         │ 
              │              ...'            │ 
            0 │       ...''''                │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
                 ndims = 3, rank = 300, mode = 1 
                ┌──────────────────────────────┐ 
          2 000 │                              │ 
                │                              │ 
                │                              │ 
                │                              │ 
Time (ms)       │                             .│ 
                │                           .' │ 
                │                         .'   │ 
                │                       .'     │ 
                │                  ...''       │ 
              0 │       .......''''            │ 
                └──────────────────────────────┘ 
                 0                          200  
                              Size               
```

</td>
<td>

```
               ndims = 3, rank = 300, mode = 2 
              ┌──────────────────────────────┐ 
          900 │                             :│ 
              │                           .' │ 
              │                          .'  │ 
              │                         :    │ 
Time (ms)     │                       .'     │ 
              │                      .'      │ 
              │                    .'        │ 
              │                 ..'          │ 
              │             ...'             │ 
            0 │       ...'''                 │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 300, mode = 3 
              ┌──────────────────────────────┐ 
          900 │                             .│ 
              │                           .' │ 
              │                          .'  │ 
              │                         :    │ 
Time (ms)     │                       .'     │ 
              │                      .'      │ 
              │                    .''       │ 
              │                  .'          │ 
              │             ...''            │ 
            0 │       ...'''                 │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
</tr>
<tr></tr>
<tr>
<th>Baseline</th>
<td>

```
               ndims = 3, rank = 50, mode = 1  
              ┌──────────────────────────────┐ 
          200 │                             :│ 
              │                            : │ 
              │                           :  │ 
              │                         .'   │ 
Time (ms)     │                        .'    │ 
              │                       .'     │ 
              │                      .'      │ 
              │                   .''        │ 
              │                .''           │ 
            0 │       ....'''''              │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 50, mode = 2  
              ┌──────────────────────────────┐ 
          200 │                              │ 
              │                              │ 
              │                             :│ 
              │                           .' │ 
Time (ms)     │                         .'   │ 
              │                        :     │ 
              │                      .'      │ 
              │                   .''        │ 
              │                .''           │ 
            0 │       ....'''''              │ 
              └──────────────────────────────┘ 
               0                          200  
                            Size               
```

</td>
<td>

```
               ndims = 3, rank = 50, mode = 3  
              ┌──────────────────────────────┐ 
          200 │                              │ 
              │                              │ 
              │                             :│ 
              │                           .' │ 
Time (ms)     │                         .'   │ 
              │                        :     │ 
              │                      .'      │ 
              │                   .''        │ 
…